storage, pass test to validate invalid input logic for the deleteDiveSite storage action

+++ src/services/AsyncStorageDataStore.ts
@@ -48,7 +48,9 @@ export class AsyncStorageDataStore implements IDataStore {
   async deleteDiveSite(id: string): Promise<void> {
     const diveSites = await this.getDiveSites();
     // Filter out the diveSite with the matching id
-    const filteredDiveSites = diveSites.filter(diveSite => diveSite.data.properties.id !== id);
+    const diveSiteToDeleteIdx = diveSites.findIndex(diveSite => diveSite.data.properties.id === id)
+    if (diveSiteToDeleteIdx === -1) throw new Error('The provided dive site ID does not exist, provide an existing ID');
+    const filteredDiveSites = diveSites.toSpliced(diveSiteToDeleteIdx, 1);
     const jsonValue = JSON.stringify(filteredDiveSites);
     await AsyncStorage.setItem(SPOTS_STORAGE_KEY, jsonValue);
   }
